Development

+++ backend/src/controller/workout.controller.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from "express"
+import { Errors } from "../error/errors"
 import { WorkoutModel } from "../model/workout.model"
 import { IWorkout } from "../types"
+import { isValidId } from "../utils/validateId"
 
 export const getAll = async (req: Request, res: Response) => {
   try {
@@ -13,6 +15,9 @@ export const getAll = async (req: Request, res: Response) => {
 
 export const getById = async (req: Request, res: Response) => {
   const { id } = req.params
+  if (!isValidId(id)) {
+    return res.status(400).json(Errors.notFound('Workout'))
+  }
   try {
     const workout = await WorkoutModel.findById(id)
     return res.status(200).json(workout)
@@ -37,8 +42,14 @@ export const deleteAll = (req: Request, res: Response) => {
 
 export const deleteById = async (req: Request, res: Response) => {
   const { id } = req.params
+  if (!isValidId(id)) {
+    return res.status(400).json(Errors.notFound('Workout'))
+  }
   try {
     const workout = await WorkoutModel.findById(id)
+    if (!workout) {
+      return res.status(404).json(Errors.notFound(`Workout ${id}`))
+    }
     await WorkoutModel.findByIdAndDelete(id)
     return res.status(200).json(workout)
   } catch (error) {
@@ -46,6 +57,21 @@ export const deleteById = async (req: Request, res: Response) => {
   }
 }
 
-export const editOne = (req: Request, res: Response) => {
-  return res.json({ msg: "editOne" })
+export const editOne = async (req: Request, res: Response) => {
+  const { id } = req.params
+  const paramsToUpdate = req.body
+
+  if (!isValidId(id)) {
+    return res.status(400).json(Errors.notFound('Workout'))
+  }
+
+  try {
+    const workout = await WorkoutModel.findOneAndUpdate({ _id: id }, { ...paramsToUpdate }, { new: true })
+    if (!workout) {
+      return res.status(404).json(Errors.notFound(`Workout ${id}`))
+    }
+    return res.status(200).json(workout)
+  } catch (error) {
+    return res.status(400).json({ error })
+  }
 }
